third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/dom/getElementById-consistency-expected.txt

CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Test that DOMSubtreeModified listeners cannot cause inDocument to be incorrect

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.contains(document.getElementById('el')) is true
PASS document.getElementById('el') is null
PASS successfullyParsed is true

TEST COMPLETE
